Predict pulling hands blocking
Think I forgor to add this one when I split the 1 morbillion PRs out.
